requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.17

alembic==1.9.4
passlib==1.7.4
python-dotenv==1.0.0

